--- src/actualServiceDeskExamples/README.md
@@ -1,3 +1,0 @@
-# Actual Service Desk Examples
-
-These examples, while functional, are examples only, and have not been vetted for production use.
