+++ src/middleware/flex/README.md
@@ -19,7 +19,7 @@ The following are some useful guides on Twilio Flex custom chat implementation:
 
 ##### Set credentials
 
-1. Copy `.env-sample` to `.env` in this directory.
+1. Copy `.env-sample` to `.env` in the current `middleware/flex` directory.
 2. Plug your credentials into `.env`
 
 You can find the following credentials in your Twilio Console:
@@ -50,7 +50,7 @@ Please note that the local server is running nonsecure on `http`. For your produ
 #### Start the webchat
 Now that the twilio middleware is up and ready, you would need to start the web-chat and start playing with the integration.
 
-* Go to the root of the project folder and copy `.env-sample` file from the same directory to `.env`. Modify the `SERVICE_DESK_CLASS` variable to `TwilioFlex`. This is the name of the class for the Twilio implementation in [twilioFlex.ts](../../serviceDesks/twilioFlex.ts). 
+* Go to the root of the project folder and copy `.env-sample` file from the same directory to `.env`. Modify the `SERVICE_DESK_CLASS` variable to `TwilioFlex`. This is the name of the class for the Twilio implementation in [twilioFlex.ts](../../serviceDesks/twilio/twilioFlex.ts). 
 * Run `npm install` 
 * Run `npm run dev`. 
 * More information on how to run the webchat can be found [here](../../../README.md#development) 
@@ -58,7 +58,7 @@ Now that the twilio middleware is up and ready, you would need to start the web-
 
 ### Functionalities
 #### Start chat with an agent
-The *startChat()* function in [twilioFlex.ts](../../serviceDesks/twilioFlex.ts) will trigger the integration with Twilio Flex. It calls the local middleware server for auth information which need to be changed according to your needs. 
+The *startChat()* function in [twilioFlex.ts](../../serviceDesks/twilio/twilioFlex.ts) will trigger the integration with Twilio Flex. It calls the local middleware server for auth information which need to be changed according to your needs. 
 
 #### End chat
 Both User and Agent ending the chat is supported. This happens by leaving/closing the twilio channel that was created when the chat begun. 
@@ -70,10 +70,10 @@ Both User and Agent can exchange text messages.
 The Twilio middleware [auth.ts](./src/routes/auth.ts) does implement a very simple user authentication by generating a user token. Please change this according to your implementation needs.
 
 #### Agent transfers
-The event listeners set up in [twilioFlex.ts](../../serviceDesks/twilioFlex.ts) listens to members joining and leaving the channel. The current implementation closes the chat if an agent leaves (`memberLeft` event), which needs to be changed for agent transfer scenario.
+The event listeners set up in [twilioFlex.ts](../../serviceDesks/twilio/twilioFlex.ts) listens to members joining and leaving the channel. The current implementation closes the chat if an agent leaves (`memberLeft` event), which needs to be changed for agent transfer scenario.
 
 #### Conversation topic
-Current implementation sends `message_to_human_agent` attribute from the dialog as the first message to the agent when the conversation starts. However, you do have access to several other attributes within the `connectMessage` object that is available in the *startChat()* function, and you can use those.
+Current implementation sends `message_to_human_agent` attribute from the dialog as the first message to the agent when the conversation starts. However, you do have access to several other attributes within the `connectMessage` object that is available in the *startChat()* function. You can also find this in our public API; Look for `topic` under `connect_to_agent` [response type](https://cloud.ibm.com/apidocs/assistant/assistant-v2#message).
 
 #### Agent name and avatar
 The agent name is implemented and available in the current implementation. The webchat does provide a method to display agent avatar/picture, but it is not implemented for Twilio Flex.
@@ -87,9 +87,8 @@ There's a crude implementation to set read receipts, but you would have to chang
 #### Agent availability
 Not implemented. Please get in touch with us, if you find a Twilio API to support this feature.
 
-
 #### Routing
-This example implementation does not take an opinion on how you should implement your contact center routing on twilio. The following documentation has information on how the routing is supported in flex and you should change the code necessary to implement the routing.
+Not Implemented. This example implementation does not take an opinion on how you should implement your contact center routing on twilio. The following documentation has information on how the routing is supported in flex and you should change the code necessary to implement the routing.
 
 * [https://www.twilio.com/docs/flex/developer/routing](https://www.twilio.com/docs/flex/developer/routing)
 * [https://www.twilio.com/docs/flex/quickstart/flex-routing-skills](https://www.twilio.com/docs/flex/developer/routing)
