+++ src/middleware/genesys/arch-overview.png
